pyproject.toml
[tool.poetry]
name = "python-sqlalchemy2"
version = "0.1.0"
description = ""
authors = ["Aleksey Eliseev <alekseyeliseev@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.32"
asyncpg = "^0.29.0"
alembic = "^1.13.2"


[tool.poetry.group.dev.dependencies]
setuptools = "^74.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
